+++ lib/src/elements/mixins/mp_bounding_box_mixin.dart
@@ -4,15 +4,15 @@ import 'package:mapiah/src/controllers/th2_file_edit_controller.dart';
 mixin MPBoundingBoxMixin {
   Rect? _boundingBox;
 
-  Rect getBoundingBox(TH2FileEditController th2FileEditController) {
+  Rect? getBoundingBox(TH2FileEditController th2FileEditController) {
     _boundingBox ??= calculateBoundingBox(th2FileEditController);
 
-    return _boundingBox!;
+    return _boundingBox;
   }
 
   void clearBoundingBox() {
     _boundingBox = null;
   }
 
-  Rect calculateBoundingBox(TH2FileEditController th2FileEditController);
+  Rect? calculateBoundingBox(TH2FileEditController th2FileEditController);
 }