+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Fixed bugs:
   * After moving a lot (more than 1.000) elements at once, finishing the move (by clicking in empty space) takes forever to complete.
   * Not showing files with raster images on web version.
+  * Unfound XVI or raster images include point 0,0 in bounding box calculation.
 * Infrastructure maintenance:
   * Centralizing controller update after element edit.